organize and write a better test
